Generate Hermes attacks in any order.
This fixes the issue that in the web UI, users had to enter the two attacks
in their canonical order, which is not very intuitive for humans, so it would
appear as if it's not possible to do attacks in some cases (i.e., when the
"wrong" target is attacked first).

This change makes the Minimax AI slightly less efficient, but that can be
fixed later.